add authorization

+++ controllers/Feed.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const path = require("path");
+const User = require("../models/User");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -15,7 +16,7 @@ exports.getPosts = (req, res, next) => {
         .then((posts) => {
           res.status(200).json({
             posts: posts,
-            totalPosts : totalItems
+            totalPosts: totalItems,
           });
         })
         .catch((err) => {
@@ -48,21 +49,34 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path;
-
+  const userId = req.userId;
   const post = new Post({
     title,
-    imageUrl: imageUrl,
     content,
-    creator: {
-      name: "Default author",
-    },
+    imageUrl: imageUrl,
+    creator: userId,
   });
+  let creator;
   post
     .save()
     .then((post) => {
+      return User.findById(userId);
+    })
+    .then((user) => {
+      if (!user) {
+        const error = new Error("User attached not found");
+        error.statusCode = 404;
+        throw error;
+      }
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully",
         post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -97,7 +111,6 @@ exports.getPost = (req, res, next) => {
 
 exports.updatePost = (req, res, next) => {
   const postId = req.params.postId;
-  console.log(postId);
   const error = validationResult(req);
   if (!error.isEmpty()) {
     const error = new Error("Validation failed, entred data is incorrect!!");
@@ -121,6 +134,11 @@ exports.updatePost = (req, res, next) => {
         console.log("not found post");
         const error = new Error("Post not found!");
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403; // not authorized
         throw error;
       }
       if (post.imageUrl !== imageUrl) {
@@ -153,8 +171,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 403; // not authorized
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({ message: "Post deleted succefully" });
